moved globus into zambeze/data dir. Moved identity to utils since can be used multiple places.
