+++ zambeze/utils/identity.py
@@ -1,3 +1,4 @@
+
 # Standard python imports
 import re
 from uuid import UUID
